requirements.txt
flask
tensorflow-cpu
keras
numpy
pandas
opencv-python
scikit-learn
pillow